Cross-validation is performed to a few models
I did apply kfold split to the female data. The baseline model is linear reg with all snps. MSE is around 36-38k. Then linreg is applied for models containing snps with lowest p-values. With the first 2000 snps, the model has mean MSE around 33k, which is better than the baseline. Then Lasso Reg is applied, and the columns are sorted with highest Lasso coefficient to the lowest. The model with the first 1570 Lasso coefficients yield mean MSE around 31k. So far, this is the best MSE I have obtained.